+++ packages/model-serving/src/shared/tracking/modelCapabilitiesTracking.ts
@@ -1,18 +1,7 @@
-import type { TrackEventFn } from './modelServingTrackingConstants';
+import { ModelServingTrackingEvent, type TrackEventFn } from './modelServingTrackingConstants';
 import { resolveWellKnownModelCapability } from '../modelCapabilities';
 
-export enum ModelCapabilitiesEvent {
-  CAPABILITY_ADDED = 'Deployment Capability Added',
-  CAPABILITY_REMOVED = 'Deployment Capability Removed',
-  CAPABILITY_MENU_OPENED = 'Deployment Capability Menu Opened',
-}
-
-export type CapabilityAddedProperties = {
-  capabilityName: string;
-  isSuggested: boolean;
-};
-
-export type CapabilityRemovedProperties = {
+export type CapabilityChangedProperties = {
   capabilityName: string;
   isSuggested: boolean;
 };
@@ -22,30 +11,37 @@ export type CapabilityMenuOpenedProperties = {
 };
 
 export type ModelDeployedCapabilityProperties = {
-  countOfCapabilities: number;
   countOfSuggestedCapabilities: number;
   countOfCustomCapabilities: number;
 };
 
+export const toCapabilityEventProps = (capability: string): CapabilityChangedProperties => {
+  const isSuggested = resolveWellKnownModelCapability(capability) != null;
+  return {
+    capabilityName: isSuggested ? capability : 'custom',
+    isSuggested,
+  };
+};
+
 export const fireCapabilityAdded = (
   trackEvent: TrackEventFn,
-  properties: CapabilityAddedProperties,
+  properties: CapabilityChangedProperties,
 ): void => {
-  trackEvent(ModelCapabilitiesEvent.CAPABILITY_ADDED, properties);
+  trackEvent(ModelServingTrackingEvent.CAPABILITY_ADDED, properties);
 };
 
 export const fireCapabilityRemoved = (
   trackEvent: TrackEventFn,
-  properties: CapabilityRemovedProperties,
+  properties: CapabilityChangedProperties,
 ): void => {
-  trackEvent(ModelCapabilitiesEvent.CAPABILITY_REMOVED, properties);
+  trackEvent(ModelServingTrackingEvent.CAPABILITY_REMOVED, properties);
 };
 
 export const fireCapabilityMenuOpened = (
   trackEvent: TrackEventFn,
   properties: CapabilityMenuOpenedProperties,
 ): void => {
-  trackEvent(ModelCapabilitiesEvent.CAPABILITY_MENU_OPENED, properties);
+  trackEvent(ModelServingTrackingEvent.CAPABILITY_MENU_OPENED, properties);
 };
 
 export const getCapabilityCounts = (capabilities: string[]): ModelDeployedCapabilityProperties => {
@@ -53,7 +49,6 @@ export const getCapabilityCounts = (capabilities: string[]): ModelDeployedCapabi
     (cap) => resolveWellKnownModelCapability(cap) != null,
   ).length;
   return {
-    countOfCapabilities: capabilities.length,
     countOfSuggestedCapabilities,
     countOfCustomCapabilities: capabilities.length - countOfSuggestedCapabilities,
   };
